Add support for no containers and no images

+++ .clean_docker.sh
@@ -3,5 +3,18 @@
 allContainersIDs=$(docker ps -aq)
 allImagesIDs=$(docker images -aq)
 
-docker rm -vf $allContainersIDs
-docker rmi -f $allImagesIDs
+if [ ${#allContainersIDs} == 0 ]
+then
+	echo "[INFO]: There is no Docker containers to remove"
+else
+	echo "[INFO]: Removing Docker containers..."
+	docker rm -vf $allContainersIDs
+fi
+
+if [ ${#allImagesIDs} == 0 ]
+then
+	echo "[INFO]: There is no Docker images to remove"
+else
+	echo "[INFO]: Removing Docker images..."
+	docker rmi -f $allImagesIDs
+fi
